refactor(scripts): consolidate timestamp normalization helpers
`normalizeApplyNow`/`toSecondPrecisionIso` were independently
re-implemented four times across idd-roadmap-audit-execute.mts,
suitability-close-execute.mts, provider-outage-park.mts, and
provider-outage-declaration.mts -- the same millisecond-precision
bug (renderUnclaimedByMarker throws on anything but second-precision
`...Z`) was caught and separately fixed twice already (#2558, #2350),
because each copy had to independently rediscover the trap.

Extract one canonical, exported, unit-tested `toSecondPrecisionIso`/
`normalizeApplyNow` pair into marker-helpers.mts, alongside
`renderUnclaimedByMarker` whose precision requirement is the reason
this normalization must exist. All four call sites now import the
shared implementation from protocol-helpers.mts (which re-exports
marker-helpers.mts) instead of defining their own private copy, and
scripts/*.mjs mirrors are regenerated via `pnpm run build`.

Direct unit tests for the shared function (millisecond truncation,
non-UTC-offset normalization to UTC, null on unparseable input) move
to tests/marker-helpers-timestamp.test.mts; the per-call-site test
files keep only their own integration/wiring regression tests.

Refs #2568
