+++ src/scripts/suitability-close-execute.mts
@@ -35,6 +35,7 @@ import {
 import { loadPolicyConfig } from './idd-config.mts';
 import type { ClaimValidationSummary } from './protocol-helpers.mts';
 import {
+  normalizeApplyNow,
   renderUnclaimedByMarker,
   resolveTrustedMarkerActors,
   summarizeClaimValidation,
@@ -190,28 +191,6 @@ export function buildSuitabilityCloseComment(evidence: string): string {
     'this close is wrong, reopen the issue -- the close is reversible and',
     'a wrong close is an accepted, recoverable risk._',
   ].join('\n');
-}
-
-/**
- * Validate and normalize the apply-time "now" to UTC second-precision ISO
- * (`YYYY-MM-DDTHH:mm:ssZ`), or `null` when unparseable. Mirrors
- * `idd-roadmap-audit-execute.mts`'s own `normalizeApplyNow` (Copilot review
- * finding on PR #2558): `deps.now()`'s production wiring is plain
- * `new Date().toISOString()`, which always carries millisecond precision,
- * but `renderUnclaimedByMarker` accepts only second-precision `…Z` and
- * throws otherwise -- and by the time `releaseClaim` runs, the evidence
- * comment and the close have already landed, so a throw here would leave
- * the issue closed with the coordination claim never released. Normalizing
- * once, fail-closed, before any mutation avoids that partial-completion
- * state entirely; the single normalized value is reused for both claim
- * re-validation and the release-marker timestamp.
- */
-export function normalizeApplyNow(raw: string): string | null {
-  const parsed = new Date(raw);
-  if (Number.isNaN(parsed.getTime())) {
-    return null;
-  }
-  return parsed.toISOString().replace(/\.\d{3}Z$/, 'Z');
 }
 
 export interface SuitabilityCloseExecuteVerdict {
